requirements.txt
langchain 
faiss-cpu 
sentence-transformers 
openai
groq 
python-dotenv
SpeechRecognition 
pyttsx3  
pyaudio
sounddevice
json5
fastapi
uvicorn[standard]